chore: move `getVersionOrDateFromTemplate` to templateUtil
