+++ src/provider/element-templates/util/templateUtil.js
@@ -1,5 +1,8 @@
 import { getBusinessObject } from 'bpmn-js/lib/util/ModelUtil';
 
+import { isUndefined } from 'min-dash';
+
+
 export function unlinkTemplate(element, injector) {
   const modeling = injector.get('modeling');
 
@@ -39,6 +42,27 @@ export function updateTemplate(element, newTemplate, injector) {
   });
 }
 
+export function getVersionOrDateFromTemplate(template) {
+  var metadata = template.metadata,
+      version = template.version;
+
+  if (metadata) {
+    if (!isUndefined(metadata.created)) {
+      return toDateString(metadata.created);
+    } else if (!isUndefined(metadata.updated)) {
+      return toDateString(metadata.updated);
+    }
+  }
+
+  if (isUndefined(version)) {
+    return null;
+  }
+
+  return version;
+}
+
+
+// helper //////
 function getEventDefinitionType(businessObject) {
   if (!businessObject.eventDefinitions) {
     return null;
@@ -51,4 +75,24 @@ function getEventDefinitionType(businessObject) {
   }
 
   return eventDefinition.$type;
+}
+
+function toDateString(timestamp) {
+  var date = new Date(timestamp);
+
+  var year = date.getFullYear();
+
+  var month = leftPad(String(date.getMonth() + 1), 2, '0');
+
+  var day = leftPad(String(date.getDate()), 2, '0');
+
+  return day + '.' + month + '.' + year;
+}
+
+function leftPad(string, length, character) {
+  while (string.length < length) {
+    string = character + string;
+  }
+
+  return string;
 }
